Add membership

+++ src/main/java/com/example/demo/dao/UserDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDao extends JpaRepository<UserEntity, Long> {
+public interface UserDao extends JpaRepository<UserEntity, Integer> {
 }
